pyproject.toml
[tool.poetry]
name = "autoria-clone"
version = "0.1.0"
description = ""
authors = ["regvud <silveron.regvud@gmail.com>"]
readme = "README.md"
packages = [{include = "autoria_clone"}]

[tool.poetry.dependencies]
python = "^3.11"
mysqlclient = "^2.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-dotenv = "^1.4.2"
django = "^4.2.7"
django-filter = "^23.3"
requests = "^2.31.0"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
redis = "^5.0.1"
daphne = "^4.0.0"
channels-redis = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
